Update MyST reference instructions to use new UX (#43)
* Update MyST reference instructions to use new UX

Instead of URL editing, the user will have features in the user
interface to enable them to build and rebuild their MyST site.

This is a compromise, not our end goal; we want the skills they use in
the JupyterLab environment to be low-frication _and_ portable to their
local machine. I.e. we want them to use `myst start`. This is currently
not possible, and the MyST team expects it may take a month or more.

* Fix header level